feat: expose diff stats manager
