+++ src/agent/output/IOutputHandler.ts
@@ -1,6 +1,6 @@
 // Local imports - types
 import { AgentStateGlobal } from '@agent/core/AgentState';
-import { NamedOutputFile } from './types';
+import { NamedOutputFile, OutputFileInfo } from './types';
 
 /** Interface describing OutputHandler behavior used by agents. */
 export interface IOutputHandler {
@@ -46,4 +46,7 @@ export interface IOutputHandler {
     currRound: number,
     groupId?: string,
   ): Promise<void>;
+
+  /** Gather mapping and diff stats for output files of a round. */
+  gatherOutputFileInfo(currRound: number): Promise<OutputFileInfo[]>;
 }
